style: Remove unused files
