--- ImageFeed/Modules/Profile/Model/Profile.swift
@@ -1,15 +1,0 @@
-//
-//  Profile.swift
-//  ImageFeed
-//
-//  Created by SERGEY SHLYAKHIN on 12.01.2023.
-//
-
-import Foundation
-
-struct Profile {
-    let image: String
-    let fullName: String
-    let loginName: String
-    let description: String
-}
